simple REST controller

+++ toolsProjects/springboot-in-10-steps/src/main/java/com/in28minutes/springboot/basics/springbootin10steps/Book.java
@@ -1,16 +1,16 @@
 package com.in28minutes.springboot.basics.springbootin10steps;
 
 public class Book {
-	int id;
+	long id;
 	String name;
 	String author;
-	public Book(int id, String name, String author) {
+	public Book(long id, String name, String author) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.author = author;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	public String getName() {
